+++ core/core.go
@@ -38,13 +38,15 @@ type ApplicationState struct {
 }
 
 type MainView struct {
-	Grid        *cview.Grid
-	Header      *cview.TextView
-	LeftMargin  *cview.TextView
-	Panels      *cview.Panels
-	StatusBar   *cview.TextView
-	PageCounter *cview.TextView
-	Settings    *cview.TextView
+	Grid          *cview.Grid
+	Header        *cview.TextView
+	LeftMargin    *cview.TextView
+	Panels        *cview.Panels
+	StatusBar     *cview.TextView
+	PageCounter   *cview.TextView
+	Settings      *cview.TextView
+	InfoScreen    *cview.TextView
+	KeymapsScreen *cview.TextView
 }
 
 type Config struct {